Add files via upload
Here are the list of pages from which code was taken to create this code:

https://github.com/adafruit/Adafruit_CircuitPython_IL0373/blob/main/examples/il0373_2.9_grayscale.py
https://learn.adafruit.com/circuitpython-display_text-library/label-updating

What this code does:

Prints serial terminal onto screen.

How this code works:

This code displays text by creating an instance of the class IL0373, which is inside of the adafruit_il0373 python file, which is a driver. Now, this driver works with the built in displayio module in the featherwing M4. Part of the reason that we have to install this driver is that our breakout board does not have the DISPLAY module, which would take care of this for us. Since we do not have the DISPLAY module, I substituted that for the driver in the code. This should allow us to use the displayio commands that are capable of displaying text onto the e paper, however, as you will see, there are some errors. I believe that these errors are fixable. 

     There is a keyboard interrupt error, which is why this code is not working. The underlying issue is unknown, although the import supervisor and disable autoreload at the end of the code were supposed to fix the issue.